requirements.txt
pyobjc-framework-ApplicationServices>=9.0
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-Quartz>=9.0
pytest>=7.0
